@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net6.0-windows7.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 		<StartupObject>IRExplorerCmd.Program</StartupObject>
 		<Platforms>AnyCPU;ARM64</Platforms>
 	</PropertyGroup>
